UpcomingWidget as frontend for FilterWidget

+++ app/src/main/java/it/niedermann/nextcloud/deck/persistence/sync/adapters/db/dao/widgets/filter/FilterWidgetDao.java
@@ -9,7 +9,7 @@ import it.niedermann.nextcloud.deck.persistence.sync.adapters.db.dao.GenericDao;
 @Dao
 public interface FilterWidgetDao extends GenericDao<FilterWidget> {
     @Query("DELETE FROM filterwidget WHERE id = :filterWidgetId")
-    void delete (Long filterWidgetId);
+    void delete (Integer filterWidgetId);
 
     @Query("SELECT * FROM FilterWidget where id = :filterWidgetId")
     FilterWidget getFilterWidgetByIdDirectly(Integer filterWidgetId);
